requirements.txt
pytest
opensimplex
textual>=0.60.0
rich>=13.0.0
pygame>=2.6.0
llama-cpp-python>=0.2.0
